+++ assets/src/ba_data/python/bastd/ui/tournamententry.py
@@ -525,7 +525,6 @@ class TournamentEntryWindow(popup.PopupWindow):
         self._launch()
 
     def _on_pay_with_ad_press(self) -> None:
-        from ba.internal import show_ad_2
 
         # If we're already entering, ignore.
         if self._entering:
@@ -547,8 +546,9 @@ class TournamentEntryWindow(popup.PopupWindow):
         cur_time = ba.time(ba.TimeType.REAL)
         if cur_time - self._last_ad_press_time > 5.0:
             self._last_ad_press_time = cur_time
-            show_ad_2('tournament_entry',
-                      on_completion_call=ba.WeakCall(self._on_ad_complete))
+            _ba.app.ads.show_ad_2('tournament_entry',
+                                  on_completion_call=ba.WeakCall(
+                                      self._on_ad_complete))
 
     def _on_ad_complete(self, actually_showed: bool) -> None: